fix: Secret<_> serialize issue when merging configs (#109)
See the code comment for an explanation of the issue:

HACK: for secret fields in the **`Config`** struct that don't follow
the `AI_GATEWAY` prefix + the double underscore separator (`__`) format.

Right now, that only applies to `HELICONE_CONTROL_PLANE_API_KEY`,
provider keys also have their own format, but **they are not fields in the
`Config` struct**.

It was intentional to allow the `HELICONE_CONTROL_PLANE_API_KEY` naming
for better clarity, as the `AI_GATEWAY__HELICONE_OBSERVABILITY__API_KEY`
version is very verbose and confusing.

The bug here is that due to the Serialize impl, when we do
`serde_json::to_value(Self::default())` above, we get the literal value
`*****` rather than the secret we want.

The fix therefore is just to re-read the value from the environment after
serializing. This is only needed to be done here in this one place, there
aren't any other functions where we merge configs like we do here.