feat: add helicone cloud config (#270)
for ease of deployment in AWS, this commit adds non-sensitive cloud
configurations for helicone. the alternative requires initcontainers or
and volume mounts and this is much simpler.

in addition, this commit updates the CI trigger rules so that we do
not have to wait for rust ci to run if we only want to update the config